# Conflicts:
#	lock.sbt
#	src/main/scala/sigmastate/eval/CostingDataContext.scala
#	src/main/scala/sigmastate/interpreter/Interpreter.scala
#	src/main/scala/sigmastate/interpreter/ProverInterpreter.scala
#	src/main/scala/sigmastate/serialization/ErgoTreeSerializer.scala
#	src/main/scala/sigmastate/serialization/transformers/AtLeastSerializer.scala
#	src/main/scala/sigmastate/utils/ByteWriter.scala
#	src/main/scala/sigmastate/utils/Extensions.scala

+++ src/test/scala/sigmastate/utxo/benchmarks/CrowdFundingContract.scala
@@ -8,7 +8,7 @@ import sigmastate.utxo.{ErgoLikeTestInterpreter, SigmaContract}
 import scala.util.Try
 
 abstract class CrowdFundingContract(
-                                     val timeout: Long,
+                                     val timeout: Int,
                                      val minToRaise: Long,
                                      val backerProver: ErgoLikeTestProvingInterpreter,
                                      val projectProver: ErgoLikeTestProvingInterpreter
